refactor: :recycle: refactorisation composant fetcher.tsx => modification nom fichier+ import du fetchDatas, import du CinemaObject provenant des interfaces
